Add script to build and export IP in Vivado
We are using a dumb way. But it works.
The script is for project-mode. The write_project_tcl command
can generate a script file, which can be used to re-build the project AS IS.

The script has so many lines, which probably are not needed.
But the benefit of using project-mode is that GUI makes our life easier:
we don't need to deal with all the scripts in shell manually.