arty_a7_system: change HLS IP to use 125MHz clock and reset
It turns out, it's those clock and reset signals that is making
the system malfunction.

I changed all the sysnet and app to use 125MHZ that is also used by
the AXI-S from Network part. Only the MC itself and the AXI interconnect
use the 83.33MHZ now.

The sysnet and app are also updated to use 8us in their scripts