app_rdma: refined HLS AXI memory access code
This commit rewrites the HLS AXI memory access logic, and significantly
improved the II and latency.

We used to use a `for` loop to go through one byte at a time.
The generated code is very bad, I think it has tons of states.
Also, more importantly, this will lead to `Implementation Failed Timing!`
(side note: the Vivado synthesis and impl takes a long time ~15-20min)

Anyhow, now the code is use a `memcpy` to do bulk transter.
And we take care of the tkeep and tlast carefully.

The takeaway is: HLS AXI coding style matters a lot. Be careful.