HLS narrow AXI write
AXI write use strobe to indicate which bytes are useful.
If the code is written in a way like `dram[0](7, 0) = data(7,0)`,
the compiler is able to identify and generate narrow AXI write.

But if the code is written in a way like this:
	dram[0](n * 8 - 1, 0) = data(n * 8 - 1 , 0)
The compiler is not able to figure about the (n*8-1) is 0x111 aligned.
So it will first generate a AXI read, then a AXI write.

Have to use a stupid idea like the code in this commit.